+++ src/agentshelf/__init__.py
@@ -1,4 +1,3 @@
 """AgentShelf package."""
 
-__version__ = "0.1.0"
-
+__version__ = "0.2.0"